MOVE server code to server.js ADD babel-register
